requirements-dev.txt
coverage==4.3.4
PyYAML==3.12
django-debug-toolbar==1.7
flake8==3.3.0
flake8_quotes==0.9.0
autopep8==1.3.1
sphinx==1.5.5
sphinx_rtd_theme==0.2.4
mock==2.0.0
